Avoid PresentationCore when pasting text on Windows
PowerShell/WMF 5.0 has a Get-Clipboard cmdlet that can be used instead.
This is still slow because launching PowerShell is slow, but it seems
like PowerShell is the only way to do this. cmd.exe and WSH have no
access to the clipboard and clip.exe can only copy, not paste.